Trait ec_gpu::GpuField

pub trait GpuField: GpuName {
    // Required methods
    fn one() -> Vec<u32>;
    fn r2() -> Vec<u32>;
    fn modulus() -> Vec<u32>;

    // Provided method
    fn sub_field_name() -> Option<String> { ... }
}

A prime field that returns the values in a representation that is suited for the use on a GPU.

Required Methods

fn one() -> Vec<u32>

Returns 1 as a vector of 32-bit limbs in little-endian non-Montgomery form (least significant limb first).

fn r2() -> Vec<u32>

Returns R ^ 2 mod P as a vector of 32-bit limbs in little-endian non-Montgomery form (least significant limb first).

fn modulus() -> Vec<u32>

Returns the field modulus as a vector of 32-bit limbs in non-Montgomery form (least significant limb first).

Provided Methods

fn sub_field_name() -> Option<String>

If the field is an extension field, then the name of the sub-field is returned.

Object Safety

This trait is not object safe.